+++ integ-test/build.gradle
@@ -12,12 +12,19 @@ repositories {
     mavenCentral()
 }
 
+configurations.all {
+    exclude group: "commons-logging", module: "commons-logging"
+}
+
 dependencies {
-    testCompile group: 'org.elasticsearch.test', name: 'framework', version: '7.6.1'
-    testCompile group: 'org.elasticsearch.client', name: 'elasticsearch-rest-high-level-client', version: '7.6.1'
+    testCompile group: 'org.elasticsearch.test', name: 'framework', version:"${es_version}"
+    testCompile group: 'org.elasticsearch.client', name: 'elasticsearch-rest-high-level-client', version:"${es_version}"
+    testCompile group: 'org.elasticsearch.client', name: 'elasticsearch-rest-client', version:"${es_version}"
     testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile group: 'org.hamcrest', name: 'hamcrest', version: '2.1'
     testCompile group: 'org.apache.logging.log4j', name: 'log4j-core', version:'2.11.1'
+    testCompile project(path: ':', configuration: 'integtest')
+    testCompile project(':plugin')
 }
 
 dependencyLicenses.enabled = false
