+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/esintgtest/TestsConstants.java
@@ -20,6 +20,9 @@ package com.amazon.opendistroforelasticsearch.sql.esintgtest;
  */
 public class TestsConstants {
 
+    public final static String PERSISTENT = "persistent";
+    public final static String TRANSIENT = "transient";
+
     public final static String TEST_INDEX = "elasticsearch-sql_test_index";
 
     public final static String TEST_INDEX_ONLINE = TEST_INDEX + "_online";
